requirements.txt
loguru==0.7.2
opentelemetry-distro==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-instrumentation-celery==0.45b0
opentelemetry-exporter-otlp==1.24.0



